web_app/requirements.txt
asgiref
certifi
Django
django-environ
django-debug-toolbar
django-summernote
djangorestframework
Pillow
psycopg2
psycopg2-binary
mysqlclient
pytz
sqlparse
gunicorn
python-memcached
easy-thumbnails
wincertstore
python-slugify
